v1.3
git-svn-id: https://plugins.svn.wordpress.org/popup-maker/trunk@1215294 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/admin/themes/metabox.php
@@ -10,7 +10,9 @@
  */
 
 // Exit if accessed directly
-if ( ! defined( 'ABSPATH' ) ) exit;
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 /** All Themes *****************************************************************/
 
@@ -22,45 +24,50 @@ if ( ! defined( 'ABSPATH' ) ) exit;
  */
 function popmake_add_popup_theme_meta_box() {
 
-	$singular = popmake_get_label_singular('popup_theme');
-	$plural = popmake_get_label_plural('popup_theme');
+	$singular = popmake_get_label_singular( 'popup_theme' );
+	$plural   = popmake_get_label_plural( 'popup_theme' );
 
 	/** Preview Window **/
-	add_meta_box( 'popmake_popup_theme_preview', __( 'Theme Preview', 'popup-maker' ),  'popmake_render_popup_theme_preview_meta_box', 'popup_theme', 'side', 'high' );
+	add_meta_box( 'popmake_popup_theme_preview', __( 'Theme Preview', 'popup-maker' ), 'popmake_render_popup_theme_preview_meta_box', 'popup_theme', 'side', 'high' );
 
 	/** Overlay Meta **/
-	add_meta_box( 'popmake_popup_theme_overlay', __( 'Overlay Settings', 'popup-maker' ),  'popmake_render_popup_theme_overlay_meta_box', 'popup_theme', 'normal', 'high' );
+	add_meta_box( 'popmake_popup_theme_overlay', __( 'Overlay Settings', 'popup-maker' ), 'popmake_render_popup_theme_overlay_meta_box', 'popup_theme', 'normal', 'high' );
 
 	/** Container Meta **/
-	add_meta_box( 'popmake_popup_theme_container', __( 'Container Settings', 'popup-maker' ),  'popmake_render_popup_theme_container_meta_box', 'popup_theme', 'normal', 'high' );
+	add_meta_box( 'popmake_popup_theme_container', __( 'Container Settings', 'popup-maker' ), 'popmake_render_popup_theme_container_meta_box', 'popup_theme', 'normal', 'high' );
 
 	/** Title Meta **/
-	add_meta_box( 'popmake_popup_theme_title', __( 'Title Settings', 'popup-maker' ),  'popmake_render_popup_theme_title_meta_box', 'popup_theme', 'normal', 'high' );
+	add_meta_box( 'popmake_popup_theme_title', __( 'Title Settings', 'popup-maker' ), 'popmake_render_popup_theme_title_meta_box', 'popup_theme', 'normal', 'high' );
 
 	/** Content Meta **/
-	add_meta_box( 'popmake_popup_theme_content', __( 'Content Settings', 'popup-maker' ),  'popmake_render_popup_theme_content_meta_box', 'popup_theme', 'normal', 'high' );
+	add_meta_box( 'popmake_popup_theme_content', __( 'Content Settings', 'popup-maker' ), 'popmake_render_popup_theme_content_meta_box', 'popup_theme', 'normal', 'high' );
 
 	/** Close Meta **/
-	add_meta_box( 'popmake_popup_theme_close', __( 'Close Settings', 'popup-maker' ),  'popmake_render_popup_theme_close_meta_box', 'popup_theme', 'normal', 'high' );
-	/** Support Meta **/
-	add_meta_box( 'popmake_popup_support', __( 'Support', 'popup-maker' ),  'popmake_render_support_meta_box', 'popup_theme', 'side', 'default' );
-	/** Share Meta **/
-	add_meta_box( 'popmake_popup_share', __( 'Share', 'popup-maker' ),  'popmake_render_share_meta_box', 'popup_theme', 'side', 'default' );
+	add_meta_box( 'popmake_popup_theme_close', __( 'Close Settings', 'popup-maker' ), 'popmake_render_popup_theme_close_meta_box', 'popup_theme', 'normal', 'high' );
 
+	if ( ! popmake_get_option( 'disable_admin_support_widget', false ) ) {
+		/** Support Meta **/
+		add_meta_box( 'popmake_popup_support', __( 'Support', 'popup-maker' ), 'popmake_render_support_meta_box', 'popup_theme', 'side', 'default' );
+	}
+	if ( ! popmake_get_option( 'disable_admin_share_widget', false ) ) {
+		/** Share Meta **/
+		add_meta_box( 'popmake_popup_share', __( 'Share', 'popup-maker' ), 'popmake_render_share_meta_box', 'popup_theme', 'side', 'default' );
+	}
 }
-add_action( 'add_meta_boxes', 'popmake_add_popup_theme_meta_box' );
 
+add_action( 'add_meta_boxes', 'popmake_add_popup_theme_meta_box' );
 
 
 function popmake_popup_theme_meta_fields() {
 	$fields = array(
 		'popup_theme_defaults_set'
 	);
-	foreach(popmake_popup_theme_meta_field_groups() as $group) {
-		foreach(apply_filters( 'popmake_popup_theme_meta_field_group_' . $group, array()) as $field) {
+	foreach ( popmake_popup_theme_meta_field_groups() as $group ) {
+		foreach ( apply_filters( 'popmake_popup_theme_meta_field_group_' . $group, array() ) as $field ) {
 			$fields[] = 'popup_theme_' . $group . '_' . $field;
 		}
 	}
+
 	return apply_filters( 'popmake_popup_theme_meta_fields', $fields );
 }
 
@@ -73,6 +80,7 @@ function popmake_popup_theme_meta_field_groups() {
 		'content',
 		'close'
 	);
+
 	return apply_filters( 'popmake_popup_theme_meta_field_groups', $groups );
 }
 
@@ -83,7 +91,8 @@ function popmake_popup_theme_meta_field_group_overlay() {
 		'background_opacity'
 	);
 }
-add_filter('popmake_popup_theme_meta_field_group_overlay', 'popmake_popup_theme_meta_field_group_overlay', 0);
+
+add_filter( 'popmake_popup_theme_meta_field_group_overlay', 'popmake_popup_theme_meta_field_group_overlay', 0 );
 
 
 function popmake_popup_theme_meta_field_group_container() {
@@ -104,7 +113,8 @@ function popmake_popup_theme_meta_field_group_container() {
 		'boxshadow_opacity',
 	);
 }
-add_filter('popmake_popup_theme_meta_field_group_container', 'popmake_popup_theme_meta_field_group_container', 0);
+
+add_filter( 'popmake_popup_theme_meta_field_group_container', 'popmake_popup_theme_meta_field_group_container', 0 );
 
 
 function popmake_popup_theme_meta_field_group_title() {
@@ -123,7 +133,8 @@ function popmake_popup_theme_meta_field_group_title() {
 		'textshadow_opacity',
 	);
 }
-add_filter('popmake_popup_theme_meta_field_group_title', 'popmake_popup_theme_meta_field_group_title', 0);
+
+add_filter( 'popmake_popup_theme_meta_field_group_title', 'popmake_popup_theme_meta_field_group_title', 0 );
 
 
 function popmake_popup_theme_meta_field_group_content() {
@@ -134,13 +145,16 @@ function popmake_popup_theme_meta_field_group_content() {
 		'font_style',
 	);
 }
-add_filter('popmake_popup_theme_meta_field_group_content', 'popmake_popup_theme_meta_field_group_content', 0);
+
+add_filter( 'popmake_popup_theme_meta_field_group_content', 'popmake_popup_theme_meta_field_group_content', 0 );
 
 
 function popmake_popup_theme_meta_field_group_close() {
 	return array(
 		'text',
 		'padding',
+		'height',
+		'width',
 		'location',
 		'position_top',
 		'position_left',
@@ -172,17 +186,17 @@ function popmake_popup_theme_meta_field_group_close() {
 		'textshadow_opacity',
 	);
 }
-add_filter('popmake_popup_theme_meta_field_group_close', 'popmake_popup_theme_meta_field_group_close', 0);
 
-
-
+add_filter( 'popmake_popup_theme_meta_field_group_close', 'popmake_popup_theme_meta_field_group_close', 0 );
 
 
 /**
  * Save post meta when the save_post action is called
  *
  * @since 1.0
+ *
  * @param int $post_id Theme (Post) ID
+ *
  * @global array $post All the data of the the current post
  * @return void
  */
@@ -196,7 +210,7 @@ function popmake_popup_theme_meta_box_save( $post_id, $post ) {
 		return;
 	}
 
-	if ( ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) || ( defined( 'DOING_AJAX') && DOING_AJAX ) || isset( $_REQUEST['bulk_edit'] ) ) {
+	if ( ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) || ( defined( 'DOING_AJAX' ) && DOING_AJAX ) || isset( $_REQUEST['bulk_edit'] ) ) {
 		return;
 	}
 
@@ -208,6 +222,29 @@ function popmake_popup_theme_meta_box_save( $post_id, $post ) {
 		return;
 	}
 
+	$field_prefix = Popmake_Popup_Theme_Fields::instance()->field_prefix;
+
+	foreach ( Popmake_Popup_Theme_Fields::instance()->get_all_fields() as $section => $fields ) {
+
+		$section_prefix = "{$field_prefix}{$section}";
+
+		$meta_values = array();
+
+		foreach ( $fields as $field => $args ) {
+
+			$field_name = "{$section_prefix}_{$field}";
+
+			if ( isset( $_POST[ $field_name ] ) ) {
+				$meta_values[ $field ] = apply_filters( 'popmake_metabox_save_' . $field_name, $_POST[ $field_name ] );
+			}
+
+		}
+
+		update_post_meta( $post_id, "popup_theme_{$section}", $meta_values );
+
+	}
+
+
 	foreach ( popmake_popup_theme_meta_fields() as $field ) {
 		if ( isset( $_POST[ $field ] ) ) {
 			$new = apply_filters( 'popmake_metabox_save_' . $field, $_POST[ $field ] );
@@ -217,36 +254,15 @@ function popmake_popup_theme_meta_box_save( $post_id, $post ) {
 		}
 	}
 
+	delete_transient( 'popmake_theme_styles' );
+
 	do_action( 'popmake_save_popup_theme', $post_id, $post );
 }
+
 add_action( 'save_post', 'popmake_popup_theme_meta_box_save', 10, 2 );
 
 
-
 /** Theme Configuration *****************************************************************/
-
-
-
-function popmake_popup_thmem_ut_extension_promotion() {
-	if(popmake_is_admin_popup_theme_page()) { ?>
-		<div id="popuptitlediv">
-			<div class="popmake-upgrade-tip">
-				<img style="" src="<?php echo POPMAKE_URL;?>/assets/images/admin/icon-unlimited-themes.png"/> <?php _e( 'Need more than one (1) theme?', 'popup-maker' ); ?> <a href="https://wppopupmaker.com/extensions/unlimited-themes?utm_source=Plugin+Admin+Theme+Editor&utm_medium=Text+Link&utm_campaign=Unlimited+Themes" target="_blank"><?php _e( 'Check out Unlimited Themes!', 'popup-maker' ); ?></a>.
-			</div>
-		</div><?php
-	}
-}
-add_action( 'edit_form_advanced', 'popmake_popup_thmem_ut_extension_promotion', 1 );
-add_action( 'edit_page_form', 'popmake_popup_thmem_ut_extension_promotion', 1 );
-
-
-
-function popmake_popup_theme_setup() {
-	if( 'popup_theme' == get_post_type() ) {
-		echo '<style type="text/css">#visibility, #favorite-actions, .add-new-h2, .tablenav, #delete-action {display:none;}</style>';
-	}
-}
-add_action( 'admin_head', 'popmake_popup_theme_setup' );
 
 /**
  * Theme Preview Metabox
@@ -259,17 +275,16 @@ add_action( 'admin_head', 'popmake_popup_theme_setup' );
  */
 function popmake_render_popup_theme_preview_meta_box() { ?>
 	<div class="empreview">
-		<div id="PopMake-Preview">
-			<div class="example-popup-overlay"></div>
-			<div class="example-popup">
-				<div class="title"><?php _e('Title Text', 'popup-maker' );?></div>
-				<div class="content"><?php do_action('popmake_example_popup_content');?></div>
-				<a class="close-popup"><?php _e('&#215;', 'popup-maker' );?></a>
-			</div>
+	<div id="PopMake-Preview">
+		<div class="example-popup-overlay"></div>
+		<div class="example-popup">
+			<div class="title"><?php _e( 'Title Text', 'popup-maker' ); ?></div>
+			<div class="content"><?php do_action( 'popmake_example_popup_content' ); ?></div>
+			<a class="close-popup"><?php _e( '&#215;', 'popup-maker' ); ?></a>
 		</div>
+	</div>
 	</div><?php
 }
-
 
 
 /**
@@ -284,13 +299,13 @@ function popmake_render_popup_theme_preview_meta_box() { ?>
 function popmake_render_popup_theme_overlay_meta_box() {
 	global $post, $popmake_options;
 	wp_nonce_field( basename( __FILE__ ), 'popmake_popup_theme_meta_box_nonce' ); ?>
-	<input type="hidden" name="popup_theme_defaults_set" value="true" />
+	<input type="hidden" name="popup_theme_defaults_set" value="true"/>
 	<div id="popmake_popup_theme_overlay_fields" class="popmake_meta_table_wrap">
-		<table class="form-table">
-			<tbody>
-				<?php do_action( 'popmake_popup_theme_overlay_meta_box_fields', $post->ID );?>
-			</tbody>
-		</table>
+	<table class="form-table">
+		<tbody>
+			<?php do_action( 'popmake_popup_theme_overlay_meta_box_fields', $post->ID ); ?>
+		</tbody>
+	</table>
 	</div><?php
 }
 
@@ -305,13 +320,13 @@ function popmake_render_popup_theme_overlay_meta_box() {
  * @return void
  */
 function popmake_render_popup_theme_container_meta_box() {
-	global $post, $popmake_options;?>
+	global $post, $popmake_options; ?>
 	<div id="popmake_popup_theme_container_fields" class="popmake_meta_table_wrap">
-		<table class="form-table">
-			<tbody>
-				<?php do_action( 'popmake_popup_theme_container_meta_box_fields', $post->ID );?>
-			</tbody>
-		</table>
+	<table class="form-table">
+		<tbody>
+			<?php do_action( 'popmake_popup_theme_container_meta_box_fields', $post->ID ); ?>
+		</tbody>
+	</table>
 	</div><?php
 }
 
@@ -326,13 +341,13 @@ function popmake_render_popup_theme_container_meta_box() {
  * @return void
  */
 function popmake_render_popup_theme_title_meta_box() {
-	global $post, $popmake_options;?>
+	global $post, $popmake_options; ?>
 	<div id="popmake_popup_theme_title_fields" class="popmake_meta_table_wrap">
-		<table class="form-table">
-			<tbody>
-				<?php do_action( 'popmake_popup_theme_title_meta_box_fields', $post->ID );?>
-			</tbody>
-		</table>
+	<table class="form-table">
+		<tbody>
+			<?php do_action( 'popmake_popup_theme_title_meta_box_fields', $post->ID ); ?>
+		</tbody>
+	</table>
 	</div><?php
 }
 
@@ -347,13 +362,13 @@ function popmake_render_popup_theme_title_meta_box() {
  * @return void
  */
 function popmake_render_popup_theme_content_meta_box() {
-	global $post, $popmake_options;?>
+	global $post, $popmake_options; ?>
 	<div id="popmake_popup_theme_content_fields" class="popmake_meta_table_wrap">
-		<table class="form-table">
-			<tbody>
-				<?php do_action( 'popmake_popup_theme_content_meta_box_fields', $post->ID );?>
-			</tbody>
-		</table>
+	<table class="form-table">
+		<tbody>
+			<?php do_action( 'popmake_popup_theme_content_meta_box_fields', $post->ID ); ?>
+		</tbody>
+	</table>
 	</div><?php
 }
 
@@ -368,13 +383,13 @@ function popmake_render_popup_theme_content_meta_box() {
  * @return void
  */
 function popmake_render_popup_theme_close_meta_box() {
-	global $post, $popmake_options;?>
+	global $post, $popmake_options; ?>
 	<div id="popmake_popup_theme_close_fields" class="popmake_meta_table_wrap">
-		<table class="form-table">
-			<tbody>
-				<?php do_action( 'popmake_popup_theme_close_meta_box_fields', $post->ID );?>
-			</tbody>
-		</table>
+	<table class="form-table">
+		<tbody>
+			<?php do_action( 'popmake_popup_theme_close_meta_box_fields', $post->ID ); ?>
+		</tbody>
+	</table>
 	</div><?php
 }
 
@@ -387,48 +402,61 @@ function popmake_render_popup_theme_close_meta_box() {
  */
 function popmake_popup_theme_post_revision_fields( $fields ) {
 	$theme_fields = popmake_popup_theme_meta_fields();
-	foreach($theme_fields as $field) {
-		$fields[$field] = __( 'Theme Overlay', ucwords( str_replace('_', ' ', str_replace('popup_theme_', '', $field) ) ), 'popup-maker' );
+	foreach ( $theme_fields as $field ) {
+		$fields[ $field ] = __( 'Theme Overlay', ucwords( str_replace( '_', ' ', str_replace( 'popup_theme_', '', $field ) ) ), 'popup-maker' );
 	}
+
 	return $fields;
 }
+
 add_filter( '_wp_post_revision_fields', 'popmake_popup_theme_post_revision_fields' );
 
 
-function popmake_popup_theme_revision_field( $value, $field, $revision) {
+function popmake_popup_theme_revision_field( $value, $field, $revision ) {
 	return get_metadata( 'post', $revision->ID, $field, true );
 }
 
 
 function popmake_add_popup_theme_revision_fields() {
-	foreach(popmake_popup_theme_meta_fields() as $field) {
+	foreach ( popmake_popup_theme_meta_fields() as $field ) {
 		add_filter( '_wp_post_revision_field_' . $field, 'popmake_popup_theme_revision_field', 10, 3 );
 	}
 }
-add_action('plugins_loaded', 'popmake_add_popup_theme_revision_fields');
+
+add_action( 'plugins_loaded', 'popmake_add_popup_theme_revision_fields' );
 
 function popmake_popup_theme_meta_restore_revision( $post_id, $revision_id ) {
 	$post = get_post( $post_id );
+	if ( $post->post_type != 'popup_theme' ) {
+		return;
+	}
 	$revision = get_post( $revision_id );
-	foreach(popmake_popup_theme_meta_fields() as $field) {
+	foreach ( popmake_popup_theme_meta_fields() as $field ) {
 		$meta = get_metadata( 'post', $revision->ID, $field, true );
-		if ( false === $meta )
+		if ( false === $meta ) {
 			delete_post_meta( $post_id, $field );
-		else
+		} else {
 			update_post_meta( $post_id, $field, $meta );
+		}
 	}
 }
+
 add_action( 'wp_restore_post_revision', 'popmake_popup_theme_meta_restore_revision', 10, 2 );
 
 function popmake_popup_theme_meta_save_revision( $post_id, $post ) {
+	if ( $post->post_type != 'popup_theme' ) {
+		return;
+	}
 	if ( $parent_id = wp_is_post_revision( $post_id ) ) {
-		foreach(popmake_popup_theme_meta_fields() as $field) {
+		foreach ( popmake_popup_theme_meta_fields() as $field ) {
 			$meta = get_post_meta( $parent_id, $field, true );
-			if ( false !== $meta )
+			if ( false !== $meta ) {
 				add_metadata( 'post', $post_id, $field, $meta );
+			}
 
 		}
 	}
 }
+
 add_action( 'save_post', 'popmake_popup_theme_meta_save_revision', 11, 2 );
